+++ config.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	flags "github.com/jessevdk/go-flags"
-
 	log "github.com/sirupsen/logrus"
+	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-crafter.git/infra"
 )
 
 // LoggerConfig logger configuration
@@ -12,9 +12,31 @@ type LoggerConfig struct {
 	Format string `long:"log-format" env:"LOG_FORMAT" default:"text" description:"Log formatter, one of text, json."`
 }
 
+// KafkaConfig is the configuration of application dealing with Kafka
+type KafkaConfig struct {
+	Address       []string `short:"k" long:"kafka-address" env:"KAFKA_ADDRESS" default:"localhost:9092" description:"Address of Kafka server to connect to"`
+	InTopic       string   `short:"i" long:"in-topic" env:"KAFKA_TOPIC_TX_CRAFTER" default:"topic-tx-crafter" description:"Kafka topic to consume message from"`
+	OutTopic      string   `short:"o" long:"out-topic" env:"KAFKA_TOPIC_TX_NONCE" default:"topic-tx-nonce" description:"Kafka topic to send message after processing"`
+	ConsumerGroup string   `short:"g" long:"consumer-group" env:"KAFKA_CRAFTER_GROUP" default:"tx-crafter-group" description:"Kafka consumer group"`
+}
+
+// WorkerConfig application configuration
+type WorkerConfig struct {
+	Slots uint `short:"w" long:"worker-slots" env:"WORKER_SLOTS" default:"100" description:"Number of messages that can be treat in parallel."`
+}
+
+// EthConfig is the configuration of application dealing with Ethereum
+type EthConfig struct {
+	URLs []string `short:"e" long:"eth-client" env:"ETH_CLIENT_URL" default:"https://ropsten.infura.io/v3/81e039ce6c8a465180822b525e3644d7"`
+}
+
 // Config worker configuration
 type Config struct {
-	Log LoggerConfig
+	Log    LoggerConfig
+	Worker WorkerConfig
+	Kafka  KafkaConfig
+	Eth    EthConfig
+	Faucet infra.FaucetConfig
 }
 
 // LoadConfig load configuration