+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ### Version 0.1.0
 
-*Unreleased*
+*Released on January 30th 2019*
+
+- [FEAT] Implement `infra.NewERC1400ABIRegistry`
+- [FEAT] Implement `infra.CreateFaucet`, `infra.SaramaCrediter`, `infra.FaucetConfig`
+- [FEAT] Implement `main.go`